Fixes a crash when loading.

+++ Source/Game/Entities/Creature/Gossip.cs
@@ -456,7 +456,7 @@ namespace Game.Misc
 
             for (int i = 0; i < SharedConst.QuestEmoteCount; ++i)
             {
-                var emote = new QuestDescEmote(quest.DetailsEmote[i], quest.DetailsEmoteDelay[i]);
+                var emote = new QuestDescEmote((int)quest.DetailsEmote[i], quest.DetailsEmoteDelay[i]);
                 packet.DescEmotes.Add(emote);
             }
 
@@ -529,7 +529,7 @@ namespace Game.Misc
             offer.SuggestedPartyMembers = quest.SuggestedPlayers;
 
             for (uint i = 0; i < SharedConst.QuestEmoteCount && quest.OfferRewardEmote[i] != 0; ++i)
-                offer.Emotes.Add(new QuestDescEmote((uint)quest.OfferRewardEmote[i], quest.OfferRewardEmoteDelay[i]));
+                offer.Emotes.Add(new QuestDescEmote(quest.OfferRewardEmote[i], quest.OfferRewardEmoteDelay[i]));
 
             offer.QuestFlags[0] = (uint)quest.Flags;
             offer.QuestFlags[1] = (uint)quest.FlagsEx;
